requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20241021
Django==5.1.4
django-ckeditor==6.7.2
django-countries==7.6.1
django-dob-widget==1.2.0
django-environ==0.11.2
django-filter==24.3
django-import-export==4.3.3
django-js-asset==3.0.1
django-phonenumber-field==8.0.0
frozenlist==1.5.0
idna==3.10
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
phonenumbers==8.13.52
pillow==11.1.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg==3.2.4
pycountry==24.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
